### 2022-10-19 #4867 @editorialbot
[PRE REVIEW]: Quilë: C++ genetic algorithms scientific library
**Submitting author:** @ttarkowski (<a href="http://orcid.org/0000-0002-6071-9389">Tomasz Tarkowski</a>)
**Repository:** https://github.com/ttarkowski/quile
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
