### 2022-10-19 #4867 @danielskatz
@ttarkowski - thanks for your submission to JOSS.

In addition, you could work on the possibly missing DOI that editorialbot suggests, but note that this may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.

Also, given that GAs are not a new topic, you might add something brief to the paper explaining how this software fits in the landscape of existing software, and why you think this will be used by others, instead of the existing software being used.
