### 2022-10-19 #4867 @ttarkowski
@danielskatz - thank you very for your service. The missing DOI suggested by @editorialbot is unfortunately incorrect and I couldn't find the proper one, therefore I plan to leave the reference without the DOI number. The *Statement of need* section in the paper already contains explanation about landscape of GAs scientific C++ libraries and information about target audience.
