### 2022-10-19 #4867 @ttarkowski
@danielskatz - I would like to suggest person with username *StefanKohekUMFERI* from the aforementioned list as a first potential reviewer of my paper. *StefanKohekUMFERI* has declared C++ and evolutionary computation as the preferred programming language and domain, respectively, which is very good fit for my paper. According to the list, *StefanKohekUMFERI* has not yet performed review for the JOSS.

I will try to suggest second potential reviewer of the paper soon.
