### 2022-10-19 #4867 @ttarkowski
@danielskatz - I would like to suggest person with username *darioizzo* from the list as a second potential reviewer. *darioizzo* has declared C++ as the preferred programming language, which is in agreement with my work. Moreover, genetic programming and artificial intelligence were declared as the preferred domains, which is close enough to the matter of my paper. According to the list, *darioizzo* has not yet performed review for the JOSS.

I hope that these two reviewer suggestions are enough for now. Please let me know if I can do more it this regard.
