### 2022-11-02 #4902 @ttarkowski
@vissarion @mbarzegary @acrlakshman @mrshu - thank you very much for your help. Please let me know if there is anything I could do to help you during the review process.
Unfortunately, the DOI number suggested by @editorialbot for "Intelligent Machinery" is not correct and I couldn't find the proper one, therefore I plan to leave that reference without the DOI number.
