### 2022-11-16 ttarkowski/quile#1 @mbarzegary
 [JOSS Review] Reproducibility and functionality
Dear author,

Generally speaking, the library works as described in the readme file of the repo, and the examples can be set up by following the instruction. However, there are minor fixes that can improve the quality of the submission:
* gnuplot is needed to run example1 and generate the output. The instruction for building/running the example can be briefly mentioned in a readme file (similar to mentioning the requirement of BGL for the mithril example). The same thing applies to example 2.
* The mithril example compiles and runs, but there is no output written to the file `evolution.dat` after 15 minutes. Maybe you need to mention any estimation of the time needed.
* The evanstar example is difficult to evaluate. I think some sample output showing the results the user will get in the end can help to understand what this example tries to do (like the ones presented for the benchmark examples).

P.S. This issue is related to https://github.com/openjournals/joss-reviews/issues/4902
