### 2022-11-16 ttarkowski/quile#2 @mbarzegary
[JOSS Review] Tests
Dear author,

The library has a set of small test cases with a Makefile intended to build and run these small test units. However, at the end of this process, it’s not clear whether the test is successful or not. As an improvement, I suggest that it reports to the user if each test passes or fails.

P.S. This issue is related to https://github.com/openjournals/joss-reviews/issues/4902
