### 2022-11-16 ttarkowski/quile#3 @mbarzegary
[JOSS Review] Summary and statement of need in the paper
Dear author,

The paper is written well, but some minor things can be improved.

* The JOSS guidelines require the summary to be “a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience”. The current summary is not appropriate for non-specialist audiences and is a bit technical. You may consider moving the functionality and theory to another section and adding a more high-level introduction to GA and the developed library.
* “The software is therefore well suited for optimization tasks with computationally expensive fitness functions, where computation time is of the order of magnitude of seconds or more, like ab initio calculations in condensed matter physics." I wonder if there is any correlation between an expensive computation and the order of magnitude of the simulation time in computational sciences. I think this sentence needs a revision since any type of simulation (like a finite element simulation) can be expensive, making the fitness function an expensive expression to evaluate.
* For the shown example usages, it can be beneficial to mention explicitly that the figures can be reproduced by running example 1 and example 2. You also need to specify how example 2 should be run to produce the results presented in figure 2 (which may need a script on the repo to reproduce this figure).
* I think it is beneficial if you mention the limitations of the library briefly.

P.S. This issue is related to https://github.com/openjournals/joss-reviews/issues/4902
