### 2022-11-16 ttarkowski/quile#4 @mbarzegary
[JOSS Review] Documentation 
Dear author,

The compiler requirement is elaborated in the tutorial file, but not enough information is provided in the readme file as the main documentation of the code. It took me some time to find out that I need GCC 10.2.1 to correctly build the examples. I first started with GCC 8 (with support for C++ 20 as described by the readme file), and then faced a couple of issues (like the `#include <concepts>` statement).

So, I suggest that you elaborate on the compiler-related things by bringing the relevant stuff from the tutorial file up to the readme file.

P.S. This issue is related to https://github.com/openjournals/joss-reviews/issues/4902
