### 2022-11-16 #4902 @mbarzegary
Hi,

I confirm the quality of the software and the effort the author has put in to develop it. Especially, I enjoyed reading the provided tutorial, which was very well written and was full of external resources and exercises as well as some cool programming tips. However, I believe some minor concerns should be addressed before accepting the submission, elaborated in a couple of issues I have opened on the software repo (listed above).

As a user with prior knowledge of how this kind of library works, reading the tutorial was the last thing I did during the review, after which I saw that some of my concerns were already addressed in the tutorial. From this point of view, some of the comments can be seen as suggestions for adding some of the information available in the tutorial to relevant sections so that users can quickly get started on the library and run the examples without going through the long, still very informative tutorial. I have pointed out a few suggestions.

I hope the author finds the comments useful.
