### 2022-11-18 ttarkowski/quile#4 @ttarkowski
Thank you very much for your feedback! The relevant information was added to the README file (cf. _List of dependencies_ section).
