### 2022-11-19 c25dc9a
Test result message (PASS or FAIL).
