### 2022-11-19 ttarkowski/quile#2 @ttarkowski
Thank you very much for your feedback! The reporting feature was added to the Makefile.
