+++ tutorial/tutorial.txt
@@ -199,7 +199,7 @@ Table of contents
 
     f(given chessboard) = - (# of queen pairs threatening each other)
 
-  The  complete  program  can  be  found   in  ``examples/example_2/``
+  The  complete  program  can  be  found   in  ``examples/example_2/''
   directory:
 
   $ less ~/repos/examples/example_2/example_2.cc
