### 2022-11-22 ttarkowski/quile#1 @ttarkowski
The relevant README file for _example_2_ was created too.
