### 2022-11-22 ttarkowski/quile#1 @ttarkowski
Information about program execution time was added to the README file of _mithril_ example.
