### 2022-11-22 ttarkowski/quile#1 @ttarkowski
Example output was added to the _evenstar_ example.
