### 2022-11-23 ttarkowski/quile#3 @ttarkowski
Thank you very much for your feedback! And thank you for your kind words! Description for the non-specialist audience was added. More detailed description was moved into separated section of the paper.
