### 2022-11-23 f4451d6
Revision of sentence (computationally expensive fitness functions).
