### 2022-11-24 511e2a2
Figure 1 - reproducibility.
