### 2022-11-24 2892806
Time complexity analysis of n queens puzzle solution.
