### 2022-11-24 ttarkowski/quile#3 @ttarkowski
Note about reproducibility of Figure 2 was added. Script for time measurement was added to the repository.
