+++ paper/paper.bib
@@ -32,6 +32,17 @@
   doi = {10.3390/economies9010006}
 }
 
+@incollection{Eshelman1993,
+  author = "Larry J. Eshelman and J. David Schaffer",
+  title = {{Real-Coded Genetic Algorithms and Interval-Schemata}},
+  year = {1993},
+  publisher = {{Elsevier}},
+  doi = {10.1016/B978-0-08-094832-4.50018-0},
+  booktitle = {{Foundations of Genetic Algorithms}},
+  volume = {2},
+  pages = {187-202}
+}
+
 @article{Gagne2006,
   author = "Christian Gagn\'e and Marc Parizeau",
   title = {{Genericity in Evolutionary Computation Software Tools: Principles and Case Study}},
