### 2022-11-24 ttarkowski/quile#3 @ttarkowski
Note about limitations of the library was added.
