### 2022-11-24 #4902 @ttarkowski
@mbarzegary - I have just closed the last issue and regenerated the paper PDF file. Please feel free to open new issues if you think that my responses are not sufficient.
