### 2022-11-30 #4902 @ttarkowski
I plan to update the _Scholarly publications and research projects using the software_ subsection soon.
