### 2022-12-09 43d6b4d
Scholarly publications and research projects using the software.
