+++ paper/paper.bib
@@ -151,12 +151,28 @@
   address = {{Geneva, Switzerland}}
 }
 
-@misc{Tarkowski2022,
+@misc{Tarkowski2022a,
   author = "Tarkowski, Tomasz",
   title = {{Genetic algorithm formulation and tuning with use of test functions}},
   year = {2022},
   publisher = {{arXiv}},
   doi = {10.48550/ARXIV.2210.03217}
+}
+
+@misc{Tarkowski2022b,
+  author = "Tarkowski, Tomasz and {Gonzalez Szwacki}, Nevill",
+  title = {{The structure of thin boron nanowires predicted using evolutionary computations}},
+  year = {2022},
+  publisher = {{arXiv}},
+  doi = {10.48550/ARXIV.2211.11901}
+}
+
+@misc{Tarkowski2022c,
+  author = "Tarkowski, Tomasz and {Gonzalez Szwacki}, Nevill",
+  title = {{Boron nanotube structure explored by evolutionary computations}},
+  year = {2022},
+  publisher = {{arXiv}},
+  doi = {10.48550/ARXIV.2212.01704}
 }
 
 @techreport{Turing1948,
