### 2022-12-09 #4902 @ttarkowski
I have just updated the _Scholarly publications and research projects using the software_ subsection and regenerated the paper PDF file. I am still waiting for one document to be published on-line, therefore I will probably perform one more update soon.
