### 2022-12-14 #4902 @ttarkowski
I have found and corrected minor and trivial bug in the library source code (missing inclusion of header file). Tests in _doc/examples_ directory are passing successfully.
