+++ doc/examples/trim.sh
@@ -1,13 +1,19 @@
 #!/bin/bash
 
-cat $1 | fold > tmp.txt
-mv tmp.txt $1
+if test -f "$1"; then
 
-lines=$(cat $1 | wc -l)
-max=50
+    cat $1 | fold > tmp.txt
+    mv tmp.txt $1
 
-if [ ${lines} -gt ${max} ]; then
-    tail -n ${max} $1 > tmp.txt
-    (echo "[Output to this point was skipped.]" ; cat tmp.txt) > $1
-    rm tmp.txt
+    lines=$(cat $1 | wc -l)
+    max=50
+
+    if [ ${lines} -gt ${max} ]; then
+	tail -n ${max} $1 > tmp.txt
+	(echo "[Output to this point was skipped.]" ; cat tmp.txt) > $1
+	rm tmp.txt
+    fi
+
 fi
+
+exit
