### 2022-12-15 #4902 @ttarkowski
I have just corrected bug in _Makefile_ in _doc/examples_ directory. The bug could caused in false negative result of the test. Tests are passing successfully after correction.
