### 2022-12-23 3ed4803
Update of the bibliography.
