+++ paper/paper.bib
@@ -161,18 +161,21 @@
 
 @misc{Tarkowski2022a,
   author = "Tarkowski, Tomasz and {Gonzalez Szwacki}, Nevill",
-  title = {{Boron nanotube structure explored by evolutionary computations}},
-  year = {2022},
-  publisher = {{arXiv}},
-  doi = {10.48550/ARXIV.2212.01704}
-}
-
-@misc{Tarkowski2022b,
-  author = "Tarkowski, Tomasz and {Gonzalez Szwacki}, Nevill",
   title = {{The structure of thin boron nanowires predicted using evolutionary computations}},
   year = {2022},
   publisher = {{arXiv}},
   doi = {10.48550/ARXIV.2211.11901}
+}
+
+@article{Tarkowski2023,
+  author = "Tarkowski, Tomasz and {Gonzalez Szwacki}, Nevill",
+  title = {{Boron Nanotube Structure Explored by Evolutionary Computations}},
+  year = {2023},
+  journal = {{Crystals}},
+  volume = {13},
+  number = {1},
+  article-number = {19},
+  doi = {10.3390/cryst13010019}
 }
 
 @techreport{Turing1948,
