### 2022-12-23 #4902 @ttarkowski
I have just replaced one preprint in bibliography with newly published article based on that preprint.
