### 2023-01-30 ttarkowski/quile#5 @vissarion
[JOSS Review] Repository references in paper
There are many references in the JOSS paper the refer to the main branch of the code repository e.g. https://github.com/ttarkowski/quile/blob/main/tutorial/tutorial.txt and https://github.com/ttarkowski/quile/blob/main/README Since the main branch could be changed after the publication of the paper that could potentially create broken links. I think it would be better to create a tagged release of the code in github (the version of that release will be referred in the paper anyways) and create (permanent) links to files of that release.
