### 2023-01-30 ttarkowski/quile#7 @vissarion
[JOSS Review] publications and projects using the software
In this section of the paper there is a mention in a search project in computational materials science (in ack the grant number is referred UMO-2016/23/B/ST3/03575) It would have been better to provide a link to that grant because it is not easy accessible through http://www.ncn.gov.pl/

Also regarding the sentence "PhD thesis (to be published on-line)", could you be more specific i.e. what is the title of the thesis, the author's name and the university. It is better to skip the note "to be published on-line" and instead create a bibliographic reference and cite that reference in the text.
