### 2023-01-30 #4902 @vissarion
I would like to thank both reviewers  @mbarzegary and @acrlakshman for their work!

@ttarkowski I have made a quick check too and created two issues https://github.com/ttarkowski/quile/issues/7 https://github.com/ttarkowski/quile/issues/5 and a PR https://github.com/ttarkowski/quile/pull/6 with minor suggestion changes for the paper. Please let me know if you are OK with those.
