### 2023-02-03 15a7533
PhD thesis bibliographic reference.
