+++ paper/paper.bib
@@ -151,7 +151,17 @@
   address = {{Geneva, Switzerland}}
 }
 
-@misc{Tarkowski2022,
+@phdthesis{Tarkowski2022,
+  author = "Tomasz Tarkowski",
+  title = {{Przewidywanie struktury krystalicznej nanodrutów z~użyciem obliczeń ewolucyjnych}},
+  year = {2022},
+  note = {{In Polish (title: ``Crystal structure prediction of nanowires using evolutionary computations''). Submitted in 2022, published on-line in 2023}},
+  school = {{Faculty of Physics, University of Warsaw}},
+  address = {{Warszawa, Poland}},
+  url = {https://depotuw.ceon.pl/bitstream/handle/item/4452/0000-DR-95841-praca.pdf?sequence=1}
+}
+
+@misc{Tarkowski2022a,
   author = "Tarkowski, Tomasz",
   title = {{Genetic algorithm formulation and tuning with use of test functions}},
   year = {2022},
@@ -159,7 +169,7 @@
   doi = {10.48550/ARXIV.2210.03217}
 }
 
-@misc{Tarkowski2022a,
+@misc{Tarkowski2022b,
   author = "Tarkowski, Tomasz and {Gonzalez Szwacki}, Nevill",
   title = {{The structure of thin boron nanowires predicted using evolutionary computations}},
   year = {2022},
