+++ paper/paper.bib
@@ -151,14 +151,12 @@
   address = {{Geneva, Switzerland}}
 }
 
-@phdthesis{Tarkowski2022,
-  author = "Tomasz Tarkowski",
-  title = {{Przewidywanie struktury krystalicznej nanodrutów z~użyciem obliczeń ewolucyjnych}},
+@misc{Tarkowski2022,
+  author = "Tarkowski, Tomasz and {Gonzalez Szwacki}, Nevill",
+  title = {{The structure of thin boron nanowires predicted using evolutionary computations}},
   year = {2022},
-  note = {{In Polish (title: ``Crystal structure prediction of nanowires using evolutionary computations''). Submitted in 2022, published on-line in 2023}},
-  school = {{Faculty of Physics, University of Warsaw}},
-  address = {{Warszawa, Poland}},
-  url = {https://depotuw.ceon.pl/bitstream/handle/item/4452/0000-DR-95841-praca.pdf?sequence=1}
+  publisher = {{arXiv}},
+  doi = {10.48550/ARXIV.2211.11901}
 }
 
 @misc{Tarkowski2022a,
@@ -169,12 +167,14 @@
   doi = {10.48550/ARXIV.2210.03217}
 }
 
-@misc{Tarkowski2022b,
-  author = "Tarkowski, Tomasz and {Gonzalez Szwacki}, Nevill",
-  title = {{The structure of thin boron nanowires predicted using evolutionary computations}},
+@phdthesis{Tarkowski2022b,
+  author = "Tomasz Tarkowski",
+  title = {{Przewidywanie struktury krystalicznej nanodrutów z~użyciem obliczeń ewolucyjnych}},
   year = {2022},
-  publisher = {{arXiv}},
-  doi = {10.48550/ARXIV.2211.11901}
+  note = {{In Polish (title: ``Crystal structure prediction of nanowires using evolutionary computations''). Submitted in 2022, published on-line in 2023}},
+  school = {{Faculty of Physics, University of Warsaw}},
+  address = {{Warszawa, Poland}},
+  url = {https://depotuw.ceon.pl/bitstream/handle/item/4452/0000-DR-95841-praca.pdf?sequence=1}
 }
 
 @article{Tarkowski2023,
