+++ paper/paper.bib
@@ -169,9 +169,8 @@
 
 @phdthesis{Tarkowski2022b,
   author = "Tomasz Tarkowski",
-  title = {{Przewidywanie struktury krystalicznej nanodrutów z~użyciem obliczeń ewolucyjnych}},
+  title = {{Przewidywanie struktury krystalicznej nanodrutów z~użyciem obliczeń ewolucyjnych (Crystal structure prediction of nanowires using evolutionary computations)}},
   year = {2022},
-  note = {{In Polish (title: ``Crystal structure prediction of nanowires using evolutionary computations''). Submitted in 2022, published on-line in 2023}},
   school = {{Faculty of Physics, University of Warsaw}},
   address = {{Warszawa, Poland}},
   url = {https://depotuw.ceon.pl/bitstream/handle/item/4452/0000-DR-95841-praca.pdf?sequence=1}
