### 2023-02-03 ttarkowski/quile#5 @ttarkowski
Thank you for your suggestions! I removed links to the software repository and replaced them with text describing locations in software archive. Tagged release v1.0.0 (meant for *software archive* paper metadata) was created and it obtained DOI number [10.5281/zenodo.7603951](https://doi.org/10.5281/zenodo.7603951). Thanks!
