### 2023-02-03 #4902 @ttarkowski
> @ttarkowski I have made a quick check too and created two issues [ttarkowski/quile#7](https://github.com/ttarkowski/quile/issues/7) [ttarkowski/quile#5](https://github.com/ttarkowski/quile/issues/5) and a PR [ttarkowski/quile#6](https://github.com/ttarkowski/quile/pull/6) with minor suggestion changes for the paper. Please let me know if you are OK with those.

@vissarion - thank you for your work! I merged PR and closed both issues after correcting the paper.
