### 2023-02-07 #4902 @vissarion
Could you please update the data of https://zenodo.org/record/7603951 such that the author and the title match exactly with those of the JOSS paper?
