+++ paper/paper.md
@@ -24,64 +24,64 @@ bibliography: paper.bib
 
 # Summary
 
-This work discusses general-purpose genetic algorithms [@Holland1975] scientific
-header-only library named Quilë. The software is written in C++20 and is
+This work discusses a general-purpose genetic algorithms [@Holland1975] scientific
+header-only library named Quilë. The software is written in C++20 and has been
 released under the terms of the MIT license. It is available at
 [https://github.com/ttarkowski/quile/](https://github.com/ttarkowski/quile/).
-The name of the library origins from fictional language Neo-Quenya and means
+The name of the library come from the fictional language Neo-Quenya and means
 "color" (cf. origin of the word *chromosome*).
 
-The genetic algorithm or, more broadly, *evolutionary computations* is a field
-of computer science devoted to the population-based, trial-and-error methods of
-problem solving. Evolutionary computations were invented by Alan Turing by
-noting, that the principles of biological evolution and genetics can be applied
-to the optimization problems [@Turing1948; @Turing1950]. One of the first
-evolutionary computations performed on a computer were done by Nils A.
+Genetic algorithms, or more broadly, *evolutionary computations*, is a field
+of computer science devoted to population-based, trial-and-error methods of
+problem solving. Evolutionary computation was invented by Alan Turing by
+noting that the principles of biological evolution and genetics can be applied
+to optimization problems [@Turing1948; @Turing1950]. One of the first
+evolutionary computations performed on a computer was done by Nils A.
 Barricelli [@Barricelli1962; @Galloway2011]. Evolutionary computations have
 found wide applications in many disciplines [@Katoch2021; @Ghaheri2015;
 @Goudos2016; @Kudjo2017; @Lee2018; @Drachal2021].
 
-The genetic algorithm is conceptually simple procedure. The aim is to find the
+The genetic algorithm is a conceptually simple procedure. The aim is to find the
 most optimal solution of a given optimization problem. First, one begins with
-some sampling of the space of potential solutions---it can be done randomly or
+some sampling of the space of potential solutions; this can be done randomly or
 *ad hoc*. This procedure forms the first population of the evolutionary process,
-i.e. its first iteration. Each candidate solution from the population is
-evaluated in terms of "fitness". "Fitter" candidate solution has greater
-probability of becoming parent and entering next population of evolution's
+i.e., its first iteration. Each candidate solution from the population is
+evaluated in terms of its "fitness". A "fitter" candidate solution has a greater
+probability of becoming a parent and entering the next population of the evolution's
 subsequent iteration. However, less "fit" individuals still can propagate,
 albeit with lower probability. This trait of evolutionary computations helps
 preventing premature optimization to the wrong local optimum. Iterations of the
-evolution continues as long as termination condition is not met. One example of
-termination condition is reaching the fitness function *plateau*.
+evolution continue as long as the termination condition is not met. One example of
+a termination condition is reaching the fitness function *plateau*.
 
 # Overview
 
-*Nature of problem:* Floating-point, integer, binary or permutation
+*Nature of problem:* Floating-point, integer, binary, or permutation
 single-objective constrained finite-dimensional optimization problems of
-arbitrary nature, i.e. maximization of $f\colon G \rightarrow \mathbb{R}$, where
+arbitrary nature, i.e., maximization of $f\colon G \rightarrow \mathbb{R}$, where
 $G \subseteq \prod_{i=0}^{c-1} X_i$, where $X_i$ is equal to set of logical
 values $\mathbb{B}$ (i.e. false and true) or is bounded subset of set of real
 numbers $\mathbb{R}$ or integer numbers $\mathbb{Z}$.
 
 *Solution method:* Single-objective constrained genetic algorithm with pure
[…]
