### 2023-02-07 #4902 @danielskatz
@ttarkowski - I'm the track editor here, and I've suggested a bunch of changes in your paper in https://github.com/ttarkowski/quile/pull/8 - please merge this, or let me know what you disagree with, then we can proceed.
