+++ paper/paper.bib
@@ -144,7 +144,7 @@
 
 @techreport{Sutton2017,
   author = "Andrew Sutton",
-  title = {Working Paper, {C}++ extensions for {C}oncepts},
+  title = {Wording Paper, {C}++ extensions for {C}oncepts},
   year = {2017},
   number = {P0734R0},
   institution = {International Organization for Standardization},
