### 2023-02-07 a85255d
Addition of several curly braces to correct typesetting of the bibliography.
