+++ paper/paper.bib
@@ -95,7 +95,7 @@
 
 @misc{Jakobovic,
   author = "Domagoj Jakobovi\'c",
-  title = {{ECF} -- Evolutionary Computation Framework},
+  title = {{ECF} -- {E}volutionary {C}omputation {F}ramework},
   url = {http://ecf.zemris.fer.hr/}
 }
 
@@ -169,7 +169,7 @@
 
 @phdthesis{Tarkowski2022b,
   author = "Tomasz Tarkowski",
-  title = {Przewidywanie struktury krystalicznej nanodrutów z~użyciem obliczeń ewolucyjnych (Crystal structure prediction of nanowires using evolutionary computations)},
+  title = {Przewidywanie struktury krystalicznej nanodrutów z~użyciem obliczeń ewolucyjnych ({C}rystal structure prediction of nanowires using evolutionary computations)},
   year = {2022},
   school = {Faculty of Physics, University of Warsaw},
   address = {Warszawa, Poland},
