### 2023-02-07 #4902 @ttarkowski
@danielskatz - the paper is almost finished. In the last commit I added several more curly braces to correct typesetting of the bibliography. Everything else looks good! Thanks!
